Moving the CosmosDB service out again
Yeah yeah... Kinda silly, but this makes sense. Let's say I want to make one specific component use WebAssembly, do I really want to have the CosmosDbService to be in the downloaded DLL files? That's just wasted space. Security isn't a problem since authentication secrets aren't in the code.

The difference now is that it just lives in the SmallsOnline.Web.Lib namespace.