Add PublicSite.Client project
This is for any WebAssembly based components. Currently I have to put these into a separate project in the current .NET 8 preview release, but maybe in the future I don't have to.

The only thing in there now is the sidebar profile image, which is for the dumb easter egg it has. lol