Convert CosmosDB responses manually
This is to take advantage of JSON source generation with the Azure CosmosDB SDK. The current versions of the SDK don't use System.Text.Json by default, but, if you override it, it's not going to use the source generated data.

This sidesteps all of that by just getting the raw response and converting it manually.